requirements.txt
streamlit==1.20.0
python-dotenv==1.0.0
google-generativeai==0.2.0
sqlite3==3.39.4
requests==2.28.2
